test: fold defer-stack tests into existing suites
